Readme update and scrape_pdf update
