+++ LICENSE.md
@@ -1,6 +1,6 @@
 BSD 3-Clause License
 
-Copyright (c) 2022, Bokulich Laboratories.
+Copyright (c) 2024, Bokulich Laboratories.
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
